:recycle: refactor: created routers for User and Auth - moved constants to environment variables
