+++ tests/test_app.py
@@ -1,8 +1,5 @@
 from http import HTTPStatus
 
-from fast_zero.schemas import UserPublic
-from fast_zero.security import create_access_token
-
 
 def test_root_deve_retornar_ok_e_ola_mundo(client):
     response = client.get('/')
@@ -11,181 +8,8 @@ def test_root_deve_retornar_ok_e_ola_mundo(client):
     assert response.json() == {'message': 'Olá Mundo!'}
 
 
-def test_exercicio_ola_mundo_em_html(client):
-    response = client.get('/exercicio-html')
+# def test_exercicio_ola_mundo_em_html(client):
+#     response = client.get('/exercicio-html')
 
-    assert response.status_code == HTTPStatus.OK
-    assert '<h1> Olá Mundo </h1>' in response.text
-
-
-def test_create_user(client):
-    response = client.post(
-        '/users/',
-        json={
-            'username': 'alice',
-            'email': 'alice@example.com',
-            'password': 'secret',
-        },
-    )
-    assert response.status_code == HTTPStatus.CREATED
-    assert response.json() == {
-        'id': 1,
-        'username': 'alice',
-        'email': 'alice@example.com',
-    }
-
-
-def test_create_user_should_return_409_username_exists(client, user):
-    response = client.post(
-        '/users/',
-        json={
-            'username': user.username,
-            'email': 'alice@example.com',
-            'password': 'secret',
-        },
-    )
-    assert response.status_code == HTTPStatus.CONFLICT
-    assert response.json() == {'detail': 'Username already exists'}
-
-
-def test_read_users(client):
-    response = client.get('/users')
-    assert response.status_code == HTTPStatus.OK
-    assert response.json() == {'users': []}
-
-
-def test_read_users_with_users(client, user):
-    user_schema = UserPublic.model_validate(user).model_dump()
-    response = client.get('/users/')
-    assert response.json() == {'users': [user_schema]}
-
-
-def test_get_user(client, user):
-    response = client.get(f'/users/{user.id}')
-
-    assert response.status_code == HTTPStatus.OK
-    assert response.json() == {
-        'username': user.username,
-        'email': user.email,
-        'id': user.id,
-    }
-
-
-def test_get_user_should_return_not_found(client):
-    response = client.get('/users/777')
-
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == {'detail': 'User not found'}
-
-
-def test_update_user(client, user, token):
-    response = client.put(
-        f'/users/{user.id}',
-        headers={'Authorization': f'Bearer {token}'},
-        json={
-            'username': 'bob',
-            'email': 'bob@example.com',
-            'password': 'mynewpassword',
-        },
-    )
-    assert response.status_code == HTTPStatus.OK
-    assert response.json() == {
-        'username': 'bob',
-        'email': 'bob@example.com',
-        'id': user.id,
-    }
-
-
-def test_update_integrity_error(client, user, token):
-    # Inserindo fausto
-    client.post(
-        '/users',
-        json={
-            'username': 'fausto',
-            'email': 'fausto@example.com',
-            'password': 'secret',
-        },
-    )
-    response_update = client.put(
-        f'/users/{user.id}',
-        headers={'Authorization': f'Bearer {token}'},
-        json={
-            'username': 'fausto',
-            'email': 'bob@example.com',
-            'password': 'mynewpassword',
-        },
-    )
-
-    assert response_update.status_code == HTTPStatus.CONFLICT
-    assert response_update.json() == {
-        'detail': 'Username or Email already exists'
-    }
-
-
-# def test_update_user_should_return_not_found(client, token):
-#     response = client.put(
-#         '/users/777',
-#         headers={'Authorization': f'Bearer {token}'},
-#         json={
-#             'username': 'bob',
-#             'email': 'bob@exemplo.com',
-#             'password': 'mynewpassword',
-#         },
-#     )
-
-    # assert response.status_code == HTTPStatus.NOT_FOUND
-    # assert response.json() == {'detail': 'User not found'}
-
-
-def test_delete_user(client, user, token):
-    response = client.delete(
-        f'/users/{user.id}',
-        headers={'Authorization': f'Bearer {token}'},
-    )
-    assert response.status_code == HTTPStatus.OK
-    assert response.json() == {'message': 'User deleted'}
-
-
-# def test_delete_user_should_return_not_found(client):
-#     response = client.delete('/users/777')
-
-#     assert response.status_code == HTTPStatus.NOT_FOUND
-#     assert response.json() == {'detail': 'User not found'}
-
-
-def test_get_current_user_not_found(client):
-    data = {'no-email': 'test'}
-    token = create_access_token(data)
-
-    response = client.delete(
-        '/users/1',
-        headers={'Authorization': f'Bearer {token}'},
-    )
-
-    assert response.status_code == HTTPStatus.UNAUTHORIZED
-    assert response.json() == {'detail': 'Could not validate credentials'}
-
-
-def test_get_current_user_does_not_exists(client):
-    data = {'sub': 'test@test'}
-    token = create_access_token(data)
-
-    response = client.delete(
-        '/users/1',
-        headers={'Authorization': f'Bearer {token}'},
-    )
-
-    assert response.status_code == HTTPStatus.UNAUTHORIZED
-    assert response.json() == {'detail': 'Could not validate credentials'}
-
-
-def test_get_token(client, user):
-    response = client.post(
-        '/token',
-        data={'username': user.email, 'password': user.clean_password},
-    )
-    token = response.json()
-
-    assert response.status_code == HTTPStatus.OK
-    assert 'access_token' in token
-    assert 'token_type' in token
+#     assert response.status_code == HTTPStatus.OK
+#     assert '<h1> Olá Mundo </h1>' in response.text
