+++ tests/test_security.py
@@ -2,17 +2,19 @@ from http import HTTPStatus
 
 from jwt import decode
 
-from fast_zero.security import SECRET_KEY, create_access_token
+from fast_zero.security import create_access_token, settings
 
 
 def test_jwt():
     data = {'test': 'test'}
     token = create_access_token(data)
 
-    decoded = decode(token, SECRET_KEY, algorithms=['HS256'])
+    result = decode(
+        token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM]
+    )
 
-    assert decoded['test'] == data['test']
-    assert 'exp' in decoded
+    assert result['test'] == data['test']
+    assert result['exp']
 
 
 def test_jwt_invalid_token(client):
